Exception handling for sections

+++ api/src/main/java/org/setup/listify/service/SectionsService.java
@@ -1,13 +1,10 @@
 package org.setup.listify.service;
 
+import org.setup.listify.exception.BadRequestException;
 import org.setup.listify.exception.ForbiddenException;
-import org.setup.listify.exception.NotFoundException;
 import org.setup.listify.model.Sections;
-import org.setup.listify.model.Tasks;
 import org.setup.listify.repo.SectionsRepository;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
 
 @Service
 public class SectionsService {
@@ -25,6 +22,9 @@ public class SectionsService {
 
     public Long createSection(Long teamLeaderID, Long projectID,
                               String sectionName, Byte sectionPosition) {
+        if (sectionName.length() > 100) {
+            throw new BadRequestException("Section name has a maximum of 100 characters");
+        }
         repository.createSection(teamLeaderID, projectID, sectionName, sectionPosition);
 
         Sections newlyCreatedSection = repository.findTopOrderBySectionIDDesc();
@@ -32,6 +32,10 @@ public class SectionsService {
     }
 
     public void updateSection(Long sectionID, Long userID, String newSectionName) {
+        if (newSectionName.length() > 100) {
+            throw new BadRequestException("Section name has a maximum of 100 characters");
+        }
+
         if (!repository.isSectionAndUserInSameProject(sectionID, userID)) {
             throw new ForbiddenException("Only sections within your project can be updated.");
         }
@@ -39,6 +43,10 @@ public class SectionsService {
     }
 
     public void deleteSectionById(Long teamLeaderID, Long sectionID) {
+        Integer userAccessToTask = repository.userHasAccessToSection(teamLeaderID, sectionID);
+        if (userAccessToTask == null || userAccessToTask == 0) {
+            throw new ForbiddenException("User does not have access to this section");
+        }
         repository.deleteSectionById(teamLeaderID, sectionID);
     }
